Cast cloud watch push period

+++ rabbitmq-cloud-watch-agent/agent.py
@@ -50,4 +50,4 @@ if __name__ == "__main__":
             os.environ.get("PUG_BROKER_AMQP_PASSWORD"),
             "/",
             os.environ.get("PUG_BROKER_AMQP_METRIC"))
-        sleep(os.getenv("PUG_BROKER_METRIC_PUSH_PERIOD", 5 * 60))
+        sleep(int(os.getenv("PUG_BROKER_METRIC_PUSH_PERIOD", 5 * 60)))
